+++ src/digestauth.js
@@ -48,7 +48,7 @@ class DigestAuth extends EventEmitter {
 
                 return data;
             } catch (error) {
-                this.emit('error', `Digest authentication error: ${error}`);
+                this.emit('error', `Digest authentication error: ${error.message ?? error}`);
             };
         };
     };