+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.3.0] - (23.08.2024)
+
+## Changes
+
+- add control over RESTFul POST JSON Object
+- fix RESRFul enable
+- add timeout to axios
+- return axios error message instead of object if exist
+
 ## [9.2.9] - (21.08.2024)
 
 ## Changes